requirements.txt
Flask==3.0.0
openai==1.7.2
python-dotenv==1.0.0
python_docx==1.1.0
Requests==2.31.0
gunicorn==21.2.0